Added related data to the database

+++ PortalAboutEverything/PortalAboutEverything.Data/Repositories/StoreRepositories.cs
@@ -1,4 +1,5 @@
-﻿using PortalAboutEverything.Data.Model;
+﻿using Microsoft.EntityFrameworkCore;
+using PortalAboutEverything.Data.Model;
 using PortalAboutEverything.Data.Model.Store;
 using System;
 using System.Collections.Generic;
@@ -11,9 +12,6 @@ namespace PortalAboutEverything.Data.Repositories
 {
     public class StoreRepositories
     {
-        private List<GoodReview> _reviews = new List<GoodReview>();
-        private int _lastReviewId = 0;
-
         private PortalDbContext _dbContext;
 
         public StoreRepositories(PortalDbContext db)
@@ -26,9 +24,21 @@ namespace PortalAboutEverything.Data.Repositories
             return _dbContext.Goods.ToList();
         }
 
-        public List<GoodReview> GetAllReviews()
+        public List<Good> GetAllGoodsWithReviews()
         {
-            return _reviews.ToList();
+            return _dbContext.Goods.Include(x => x.Reviews).ToList();
+        }
+
+        public Good GetGoodByIdWithReview(int id)
+        {
+            var goodById = _dbContext.Goods.Include(x => x.Reviews).FirstOrDefault(x => x.Id == id);
+            return goodById;
+        }
+
+        public Good GetGoodById(int id)
+        {
+            var goodById = _dbContext.Goods.FirstOrDefault(x => x.Id == id);
+            return goodById;
         }
 
         public void AddGood(Good good)
@@ -37,29 +47,40 @@ namespace PortalAboutEverything.Data.Repositories
             _dbContext.SaveChanges();
         }
 
+        //public void Delete(int id)
+        //{
+        //    var good = _dbContext.Goods.First(x => x.Id == id);
+        //    _dbContext.Goods.Remove(good);
+        //    _dbContext.SaveChanges();
+        //}
+
         public void Delete(int id)
         {
-            var good = _dbContext.Goods.First(x => x.Id == id);
-            _dbContext.Goods.Remove(good);
-            _dbContext.SaveChanges();
+            var good = _dbContext.Goods.Include(x => x.Reviews).FirstOrDefault(x => x.Id == id);
+
+            if (good != null)
+            {
+                _dbContext.Goods.Remove(good);
+                _dbContext.SaveChanges();
+            }
         }
 
         public Good GetGoodForUpdate(int id)
         {
             return _dbContext.Goods.Single(x => x.Id == id);
-           
+
         }
 
         public void UpdateGood(Good good)
         {
-            Delete(good.Id);
-            AddGood(good);
-        }
+            var dbGood = GetGoodById(good.Id);
 
-        public void AddReview(GoodReview review)
-        {
-            review.Id = _lastReviewId++;
-            _reviews.Add(review);
+
+            dbGood.Name = good.Name;
+            dbGood.Description = good.Description;
+            dbGood.Price = good.Price;
+
+            _dbContext.SaveChanges();
         }
     }
 }
